+++ trove/derive/osfmap_json.py
@@ -1,7 +1,6 @@
 import json
 
-from trove.render.jsonld import RdfJsonldRenderer
-from trove.vocab.osfmap import OSFMAP_VOCAB, osfmap_labeler
+from trove.render.osfmap_jsonld import RdfOsfmapJsonldRenderer
 from trove.vocab.trove import TROVE
 from ._base import IndexcardDeriver
 
@@ -19,7 +18,7 @@ class OsfmapJsonDeriver(IndexcardDeriver):
     # abstract method from IndexcardDeriver
     def derive_card_as_text(self):
         return json.dumps(
-            RdfJsonldRenderer(OSFMAP_VOCAB, osfmap_labeler).tripledict_as_nested_jsonld(
+            RdfOsfmapJsonldRenderer().tripledict_as_nested_jsonld(
                 self.data.tripledict,
                 self.focus_iri,
             )
