+++ trove/views/browse.py
@@ -12,25 +12,39 @@ from trove.vocab.namespaces import TROVE
 
 
 class BrowseIriView(View):
-    def get(self, request, iri):
-        _iri = unquote_iri(iri)
+    def get(self, request, **kwargs):
+        _iri_param = kwargs.get('iri') or request.GET.get('iri')
+        if not _iri_param:
+            raise ValueError('TODO: random browse?')
+        _iri = unquote_iri(_iri_param)
         _trove_term = _recognize_trove_term(_iri)
         if _trove_term is not None:
             return redirect('trove-vocab', vocab_term=_trove_term)
+        _focus_iri, _tripledict = self._get_rdf(_iri)
+        return get_renderer(request).render_response(
+            _tripledict,
+            _focus_iri,
+            headers={
+                'Content-Disposition': 'inline',
+            },
+        )
+
+    def _get_rdf(self, iri: str):
         try:
-            _identifier = trove_db.ResourceIdentifier.objects.get_for_iri(_iri)
+            _identifier = trove_db.ResourceIdentifier.objects.get_for_iri(iri)
         except trove_db.ResourceIdentifier.DoesNotExist:
             raise http.Http404
         _rdf_qs = (
             trove_db.LatestIndexcardRdf.objects
             .filter(indexcard__focus_identifier_set=_identifier)
         )
-        # TODO: query param for split/combined
-        _tripledict = _merge_tripledicts(
+        # TODO: handle different focus_iri on multiple cards
+        _focus_iri = _rdf_qs.first().focus_iri
+        # TODO: query param for split/merged
+        return _focus_iri, _merge_tripledicts(
             _indexcard_rdf.as_rdf_tripledict()
             for _indexcard_rdf in _rdf_qs
         )
-        return get_renderer(request).render_response(_tripledict, _iri)
 
 
 def _merge_tripledicts(tripledicts: Iterable[dict]):
